refactor(Controller): Moving functions into separate fils
