+++ src/plugin/controller.ts
@@ -1,6 +1,7 @@
 import { createTable } from "./table-creator/tableCreator";
 import { sendCurrentTextSelection } from "./language-linter/languageLinter";
 import { getColorStats, getColorTokens } from "./color-linter/colorLinter";
+import { switchToTheme } from "./theme-switcher/themeSwitcher";
 
 import { selectAndZoomToLayer } from "../app/components/utils";
 
@@ -10,154 +11,11 @@ let uiSize = {
   height: 448,
 };
 
-// ==============================================================
-// Theme switcher
-// ==============================================================
-
-// Define the notification here so we can cancel it later
-let themeSwitchedNotification = undefined;
-
-const switchToTheme = async (
-  theme: "light" | "dark",
-  closeAfterRun: Boolean = false
-) => {
-  if (closeAfterRun) {
-    figma.showUI(__html__, { width: 70, height: 0 });
-  } else {
-    figma.ui.postMessage({ type: `loading-${theme}-theme-switch` });
-  }
-
-  // Check for a selection. If none exists, show error notification.
-  if (figma.currentPage.selection.length === 0) {
-    // If the notification is already set, turn it off
-    themeSwitchedNotification && themeSwitchedNotification?.cancel();
-
-    figma.notify("Select some layers before choosing a theme", { error: true });
-    return closeAfterRun && figma.closePlugin();
-  }
-
-  // If the notification is already set, turn it off
-  themeSwitchedNotification && themeSwitchedNotification?.cancel();
-
-  // Tell the user we're working on the theme change
-  const loadingNotification = figma.notify(
-    `Converting selection to ${theme} mode...`
-  );
-
-  // Get the list of colors that are using one core color styles
-  const colorStats = await getColorStats(true);
-
-  let importedStyles = [];
-  let keysToLoad: () => string[] = () => {
-    let keys = [];
-
-    colorStats.colorsUsingOneCoreStyle.forEach((color) => {
-      if ("theme" in color.token && color.token?.theme !== theme) {
-        const keyOfTokenInOppositeTheme =
-          theme === "light"
-            ? color.token.lightThemeToken
-            : color.token.darkThemeToken;
-        const keyIsNotDuplicate = !keys.some(
-          (key) => key === keyOfTokenInOppositeTheme
-        );
-
-        // Check to see if there's an available token to switch to
-        if (keyOfTokenInOppositeTheme === null) {
-          console.error(
-            `Missing token: No ${theme} theme token found for "${color.token.name}".`
-          );
-          return;
-        }
-
-        // Add it to the keys array if it's not already there
-        keyIsNotDuplicate && keys.push(keyOfTokenInOppositeTheme);
-      }
-    });
-
-    return keys;
-  };
-
-  // Fetch the tokens
-  importedStyles = await Promise.all(
-    keysToLoad().map(async (key) => figma.importStyleByKeyAsync(key))
-  );
-
-  // Replace every one core color style with it's
-  // dark mode equivalent
-  for (const color of colorStats.colorsUsingOneCoreStyle) {
-    if ("theme" in color.token && color.token?.theme !== theme) {
-      const node = figma.getNodeById(color.layerId);
-      const keyOfTokenInOppositeTheme =
-        theme === "light"
-          ? color.token.lightThemeToken
-          : color.token.darkThemeToken;
-
-      const importedStyle: BaseStyle = importedStyles.filter(
-        (style) => style.key === keyOfTokenInOppositeTheme
-      )[0];
-
-      // Check to see if there's an available token to switch to
-      if (keyOfTokenInOppositeTheme === null) {
-        console.error(
-          `Missing token: No ${theme} theme token found for "${color.token.name}".`
-        );
-        return;
-      }
-
-      // Get ready to set the token on the layer
-      // First, check to see if the layer has segment styles
-      if (color.layerHasSegmentStyles) {
-        (node as TextNode).setRangeFillStyleId(
-          color.segment.start,
-          color.segment.end,
-          importedStyle.id
-        );
-      } else {
-        // Set the token on the layer
-        node[`${color.colorType}StyleId`] = importedStyle.id;
-      }
-    }
-  }
-
-  loadingNotification.cancel();
-
-  const coverageAsInteger = colorStats.oneCoreColorStyleCoverage.substring(
-    0,
-    colorStats.oneCoreColorStyleCoverage.length - 1
-  );
-  const isErrorWorthy = parseInt(coverageAsInteger) < 50;
-
-  if (colorStats.oneCoreColorStyleCoverage === "100%") {
-    themeSwitchedNotification = figma.notify(
-      `${theme === "light" ? "🔆" : "🌙"} Selection set to ${theme} mode`
-    );
-  } else {
-    themeSwitchedNotification = figma.notify(
-      `✋ Warning: Only ${colorStats.oneCoreColorStyleCoverage} converted to 
-      ${theme} mode because some colors aren't using One Core color styles.`,
-      { timeout: isErrorWorthy ? 999999999 : 15000, error: isErrorWorthy }
-    );
-  }
-
-  figma.ui.postMessage({
-    type: "theme-switched",
-    message: {
-      switchedTo: theme,
-      closeAfterRun,
-      colorsUsingOneCoreStyle: colorStats.colorsUsingOneCoreStyle.length,
-      colorsSelected: colorStats.allInstancesOfColor.length,
-      colorsSwitched: colorStats.oneCoreColorStyleCoverage,
-      ...customEventData,
-    },
-  });
-};
-
-// ==============================================================
-// Handle navigation
-// ==============================================================
-
-// Used to send a custom event to New Relic
-
+// Prepare custom meta data to send to New Relic
+//
+// TODO: Move this into it's own file to be imported everywhere
+//       it's used. Rather than adding an extra argument in each
+//       function where it's used (the current implementation).
 const fileName = encodeURIComponent(figma.currentPage.parent.name);
 const currentSelection = figma.currentPage.selection;
 const currentNodeId = encodeURIComponent(
@@ -178,6 +36,10 @@ let customEventData = {
   "Session ID": figma.currentUser.sessionId,
   fileUrl,
 };
+
+// ==============================================================
+// Handle navigation
+// ==============================================================
 
 // handle submenu navigation
 const navigateTo = (screen: string) => {
@@ -200,10 +62,10 @@ const navigationActions = {
     navigateTo("open-table-creator");
   },
   "theme-switcher-to-light": () => {
-    switchToTheme("light", true);
+    switchToTheme("light", true, customEventData);
   },
   "theme-switcher-to-dark": () => {
-    switchToTheme("dark", true);
+    switchToTheme("dark", true, customEventData);
   },
   "open-language-linter": () => {
     figma.showUI(__html__, { themeColors: true, width: 475, height: 500 });
@@ -215,6 +77,7 @@ const navigationActions = {
   },
 };
 
+// Choose the proper route based on the incoming command
 navigationActions[figma.command]();
 
 // ==============================================================
@@ -407,10 +270,10 @@ figma.ui.onmessage = async (msg) => {
 
   /*-- Theme switcher messages --*/
   if (msg.type === "theme-switcher-to-dark") {
-    switchToTheme("dark");
+    switchToTheme("dark", false, customEventData);
   }
   if (msg.type === "theme-switcher-to-light") {
-    switchToTheme("light");
+    switchToTheme("light", false, customEventData);
   }
 };
